separate existing code into separate clients and allow for mocking of individual clients
