+++ backend/src/Altinn.AccessManagement.UI/Altinn.AccessManagement.UI.Integration/Clients/AccessManagementClient.cs
@@ -116,173 +116,6 @@ namespace Altinn.AccessManagement.UI.Integration.Clients
         }
 
         /// <inheritdoc />
-        public async Task<HttpResponseMessage> ClearAccessCacheOnRecipient(string party, BaseAttribute recipient)
-        {
-            string endpointUrl = $"internal/{party}/accesscache/clear";
-            string token = JwtTokenUtil.GetTokenFromContext(_httpContextAccessor.HttpContext, _platformSettings.JwtCookieName);
-            StringContent requestBody = new StringContent(JsonSerializer.Serialize(recipient, _serializerOptions), Encoding.UTF8, "application/json");
-            HttpResponseMessage response = await _client.PutAsync(token, endpointUrl, requestBody);
-            return response;
-        }
-
-        //// MaskinportenSchema
-
-        /// <inheritdoc />
-        public async Task<List<MaskinportenSchemaDelegation>> GetReceivedMaskinportenSchemaDelegations(string party)
-        {
-            try
-            {
-                string endpointUrl = $"{party}/maskinportenschema/received";
-                string token = JwtTokenUtil.GetTokenFromContext(_httpContextAccessor.HttpContext, _platformSettings.JwtCookieName);
-
-                HttpResponseMessage response = await _client.GetAsync(token, endpointUrl);
-
-                if (response.StatusCode == HttpStatusCode.OK)
-                {
-                    string responseContent = await response.Content.ReadAsStringAsync();
-                    List<MaskinportenSchemaDelegation> inboundDelegations = JsonSerializer.Deserialize<List<MaskinportenSchemaDelegation>>(responseContent, _serializerOptions);
-                    return inboundDelegations;
-                }
-
-                _logger.LogError("Getting received delegations from accessmanagement failed with {StatusCode}", response.StatusCode);
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, "AccessManagement.UI // MaskinportenSchemaClient // GetInboundDelegations // Exception");
-                throw;
-            }
-
-            return null;
-        }
-
-        /// <inheritdoc />
-        public async Task<List<MaskinportenSchemaDelegation>> GetOfferedMaskinportenSchemaDelegations(string party)
-        {
-            try
-            {
-                string endpointUrl = $"{party}/maskinportenschema/offered";
-                string token = JwtTokenUtil.GetTokenFromContext(_httpContextAccessor.HttpContext, _platformSettings.JwtCookieName);
-
-                HttpResponseMessage response = await _client.GetAsync(token, endpointUrl);
-
-                if (response.StatusCode == HttpStatusCode.OK)
-                {
-                    string responseContent = await response.Content.ReadAsStringAsync();
-                    List<MaskinportenSchemaDelegation> outboundDelegations = JsonSerializer.Deserialize<List<MaskinportenSchemaDelegation>>(responseContent, _serializerOptions);
-                    return outboundDelegations;
-                }
-
-                _logger.LogError("Getting offered delegations from accessmanagement failed with {StatusCode}", response.StatusCode);
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, "AccessManagement.UI // DelegationsClient // GetOutboundDelegations // Exception");
-                throw;
-            }
-
-            return null;
-        }
-
-        /// <inheritdoc />
-        public async Task<HttpResponseMessage> RevokeReceivedMaskinportenScopeDelegation(string party, RevokeReceivedDelegation delegation)
-        {
-            try
-            {
-                string endpointUrl = $"{party}/maskinportenschema/received/revoke";
-                string token = JwtTokenUtil.GetTokenFromContext(_httpContextAccessor.HttpContext, _platformSettings.JwtCookieName);
-                StringContent requestBody = new StringContent(JsonSerializer.Serialize(delegation, _serializerOptions), Encoding.UTF8, "application/json");
-                HttpResponseMessage response = await _client.PostAsync(token, endpointUrl, requestBody);
-                return response;
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, "AccessManagement.UI // DelegationsClient // RevokeReceivedMaskinportenScopeDelegation // Exception");
-                throw;
-            }
-        }
-
-        /// <inheritdoc />
-        public async Task<HttpResponseMessage> RevokeOfferedMaskinportenScopeDelegation(string party, RevokeOfferedDelegation delegation)
-        {
-            string endpointUrl = $"{party}/maskinportenschema/offered/revoke";
-            string token = JwtTokenUtil.GetTokenFromContext(_httpContextAccessor.HttpContext, _platformSettings.JwtCookieName);
-            StringContent requestBody = new StringContent(JsonSerializer.Serialize(delegation, _serializerOptions), Encoding.UTF8, "application/json");
-            HttpResponseMessage response = await _client.PostAsync(token, endpointUrl, requestBody);
-            return response;
-        }
-
-        /// <inheritdoc />
-        public async Task<HttpResponseMessage> CreateMaskinportenScopeDelegation(string party, DelegationInput delegation)
-        {
-            string endpointUrl = $"{party}/maskinportenschema/offered";
-            string token = JwtTokenUtil.GetTokenFromContext(_httpContextAccessor.HttpContext, _platformSettings.JwtCookieName);
-            StringContent requestBody = new StringContent(JsonSerializer.Serialize(delegation, _serializerOptions), Encoding.UTF8, "application/json");
-            HttpResponseMessage response = await _client.PostAsync(token, endpointUrl, requestBody);
-            return response;
-        }
-
-        /// <inheritdoc />
-        public async Task<List<DelegationResponseData>> MaskinportenSchemaDelegationCheck(string partyId, Right request)
-        {
-            try
-            {
-                string endpointUrl = $"{partyId}/maskinportenschema/delegationcheck";
-                string token = JwtTokenUtil.GetTokenFromContext(_httpContextAccessor.HttpContext, _platformSettings.JwtCookieName);
-                StringContent requestBody = new StringContent(JsonSerializer.Serialize(request, _serializerOptions), Encoding.UTF8, "application/json");
-                HttpResponseMessage response = await _client.PostAsync(token, endpointUrl, requestBody);
-
-                if (response.StatusCode == HttpStatusCode.OK)
-                {
-                    string responseContent = await response.Content.ReadAsStringAsync();
-                    return JsonSerializer.Deserialize<List<DelegationResponseData>>(responseContent, _serializerOptions);
-                }
-                else
-                {
-                    string responseContent = await response.Content.ReadAsStringAsync();
-                    HttpStatusException error = JsonSerializer.Deserialize<HttpStatusException>(responseContent, _serializerOptions);
-
-                    throw error;
-                }
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, "AccessManagement.UI // MaskinportenSchemaClient // DelegationCheck // Exception");
-                throw;
-            }
-        }
-
-        //// SingleRights
-
-        /// <inheritdoc />
-        public async Task<HttpResponseMessage> CheckSingleRightsDelegationAccess(string partyId, Right request)
-        {
-            try
-            {
-                string endpointUrl = $"internal/{partyId}/rights/delegation/delegationcheck";
-                string token = JwtTokenUtil.GetTokenFromContext(_httpContextAccessor.HttpContext, _platformSettings.JwtCookieName);
-                StringContent requestBody = new StringContent(JsonSerializer.Serialize(request, _serializerOptions), Encoding.UTF8, "application/json");
-                HttpResponseMessage response = await _client.PostAsync(token, endpointUrl, requestBody);
-
-                return response;
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, "AccessManagement.UI // SingleRightClient // CheckDelegationAccess // Exception");
-                throw;
-            }
-        }
-
-        /// <inheritdoc />
-        public async Task<HttpResponseMessage> CreateSingleRightsDelegation(string party, DelegationInput delegation)
-        {
-            string endpointUrl = $"internal/{party}/rights/delegation/offered";
-            string token = JwtTokenUtil.GetTokenFromContext(_httpContextAccessor.HttpContext, _platformSettings.JwtCookieName);
-            StringContent requestBody = new StringContent(JsonSerializer.Serialize(delegation, _serializerOptions), Encoding.UTF8, "application/json");
-            HttpResponseMessage response = await _client.PostAsync(token, endpointUrl, requestBody);
-            return response;
-        }
-
-        /// <inheritdoc />
         public async Task<HttpResponseMessage> GetSingleRightsForRightholder(string party, string userId)
         {
             string endpointUrl = $"todo/{party}/{userId}"; // TODO: Switch with actual backend endpoint when available
