--- backend/src/Altinn.AccessManagement.UI/Altinn.AccessManagement.UI.Mocks/Data/ResourceRegistry/nav_aa_distribution/resource.json
@@ -1,62 +1,0 @@
-{
-  "title": {
-    "en": "NAV aa distribution",
-    "nb": "NAV aa distribution",
-    "nn": "NAV aa distribution"
-  },
-  "sector": [
-    "privat"
-  ],
-  "status": "Active",
-  "validTo": "2022-12-24T23:59:59",
-  "homepage": "https://www.nav.no/",
-  "isPartOf": "navportal",
-  "keywords": [
-    {
-      "word": "delegation",
-      "language": "nb"
-    }
-  ],
-  "validFrom": "2022-09-20T06:46:07.598Z",
-  "identifier": "nav_aa_distribution",
-  "isComplete": true,
-  "description": {
-    "en": "NAVs brukerstøtte for Aa-registerets distribusjonsløsninger",
-    "nb": "NAVs brukerstøtte for Aa-registerets distribusjonsløsninger",
-    "nn": "NAVs brukerstøtte for Aa-registerets distribusjonsløsninger"
-  },
-  "resourceType": "MaskinportenSchema",
-  "thematicArea": "http://publications.europa.eu/resource/authority/eurovoc/2468",
-  "isPublicService": true,
-  "rightDescription": {
-    "en": "Gir mottaker lov til å bruke api",
-    "nb": "Gir mottaker lov til å bruke api",
-    "nn": "Gir mottaker lov til å bruke api"
-  },
-  "resourceReferences": [
-    {
-      "reference": "urn:altinn:resource:nav_aa",
-      "referenceType": "Uri",
-      "referenceSource": "Altinn3"
-    },
-    {
-      "reference": "urn:altinn:maskinporten:nav_aa",
-      "referenceType": "MaskinportenScope",
-      "referenceSource": "Altinn3"
-    },
-    {
-      "reference": "altinn:test/theworld.write",
-      "referenceType": "MaskinportenScope",
-      "referenceSource": "Altinn3"
-    }
-  ],
-  "hasCompetentAuthority": {
-    "orgcode": "NAV",
-    "organization": "991825827",
-    "name": {
-      "en": "The Norwegian Labour and Welfare Administration",
-      "nb": "Arbeids og velferdsetaten",
-      "nn": "Arbeids og velferdsetaten"
-    }
-  }
-}
