+++ backend/src/Altinn.AccessManagement.UI/Altinn.AccessManagement.UI.Mocks/Data/ResourceRegistry/ske-innrapportering-boligsameie/resource.json
@@ -1,0 +1,62 @@
+{
+  "identifier": "ske-innrapportering-boligsameie",
+  "version": "1",
+  "title": {
+    "en": "Submission Jointly owned housing property",
+    "nb": "Innrapportering Boligsameie",
+    "nn": "Innrapportering Bustadsameige"
+  },
+  "description": {
+    "en": "Jointly owned housing property must submit information to the Directorate of Taxes about the unit owner/shareholder's share of any income, expenses, assets and liabilities.",
+    "nb": "Sameier med 9 eller flere boligseksjoner skal levere opplysninger om sameiers andel av felles inntekter, utgifter, formue og gjeld. Det er vanligvis sameiets styre eller forretningsfører som leverer opplysningene.",
+    "nn": "Sameier med 9 eller flere boligseksjoner skal levere opplysninger om sameiers andel av felles inntekter, utgifter, formue og gjeld. Det er vanligvis sameiets styre eller forretningsfører som leverer opplysningene."
+  },
+  "rightDescription": {
+    "en": "Provide access to using the form for submitting information on Jointly owned housing property",
+    "nb": "Delegering av denne tjenesten gir andre mulighet til å rapportere opplysninger om boligsameie på vegne av deg.",
+    "nn": "Delegering av denne tjenesten gir andre mulighet til å rapportere opplysninger om bustadsameige på vegne av deg."
+  },
+  "status": "Completed",
+  "contactPoints": [
+    {
+      "category": "Innrapportering",
+      "email": "",
+      "telephone": "",
+      "contactPage": "www.skatteetaten.no"
+    }
+  ],
+  "delegable": true,
+  "visible": true,
+  "hasCompetentAuthority": {
+    "name": {
+      "en": "Norwegian Tax Administration",
+      "nb": "Skatteetaten",
+      "nn": "Skatteetaten"
+    },
+    "organization": "974761076",
+    "orgcode": "skd"
+  },
+  "keywords": [
+    {
+      "word": "Boligsameie",
+      "language": "nb"
+    },
+    {
+      "word": "innrapportering",
+      "language": "nb"
+    }
+  ],
+  "accessListMode": "Disabled",
+  "selfIdentifiedUserEnabled": false,
+  "enterpriseUserEnabled": true,
+  "resourceType": "GenericAccessResource",
+  "availableForType": [
+    "LegalEntityEnterprise"
+  ],
+  "authorizationReference": [
+    {
+      "id": "urn:altinn:resource",
+      "value": "ske-innrapportering-boligsameie"
+    }
+  ]
+}
