+++ backend/src/Altinn.AccessManagement.UI/Altinn.AccessManagement.UI.Mocks/Mocks/AccessManagementClientMock.cs
@@ -102,173 +102,6 @@ namespace Altinn.AccessManagement.UI.Mocks.Mocks
         }
 
         /// <inheritdoc />
-        public Task<HttpResponseMessage> ClearAccessCacheOnRecipient(string party, BaseAttribute recipient)
-        {
-            return Task.FromResult(new HttpResponseMessage
-            { StatusCode = HttpStatusCode.OK });
-        }
-
-        //// MaskinportenSchema
-
-        public Task<List<MaskinportenSchemaDelegation>> GetReceivedMaskinportenSchemaDelegations(string party)
-        {
-            ThrowExceptionIfTriggerParty(party);
-
-            List<MaskinportenSchemaDelegation> delegations = new List<MaskinportenSchemaDelegation>();
-            List<MaskinportenSchemaDelegation> filteredDelegations = new List<MaskinportenSchemaDelegation>();
-
-            string path = Path.Combine(dataFolder, "MaskinportenSchema");
-            if (Directory.Exists(path))
-            {
-                string content = File.ReadAllText(Path.Combine(path, "backendReceived.json"));
-                delegations = JsonSerializer.Deserialize<List<MaskinportenSchemaDelegation>>(content, options) ?? [];
-
-                if (!string.IsNullOrEmpty(party))
-                {
-                    filteredDelegations.AddRange(delegations.FindAll(od => od.CoveredByPartyId == Convert.ToInt32(party)));
-                }
-            }
-
-            return Task.FromResult(filteredDelegations);
-        }
-
-        public Task<List<MaskinportenSchemaDelegation>> GetOfferedMaskinportenSchemaDelegations(string party)
-        {
-            ThrowExceptionIfTriggerParty(party);
-
-            List<MaskinportenSchemaDelegation> delegations = new List<MaskinportenSchemaDelegation>();
-            List<MaskinportenSchemaDelegation> filteredDelegations = new List<MaskinportenSchemaDelegation>();
-
-            string path = Path.Combine(dataFolder, "MaskinportenSchema");
-            if (Directory.Exists(path))
-            {
-                string content = File.ReadAllText(Path.Combine(path, "backendOffered.json"));
-                delegations = JsonSerializer.Deserialize<List<MaskinportenSchemaDelegation>>(content, options) ?? [];
-
-                if (!string.IsNullOrEmpty(party))
-                {
-                    filteredDelegations.AddRange(delegations.FindAll(od => od.OfferedByPartyId == Convert.ToInt32(party)));
-                }
-            }
-
-            return Task.FromResult(filteredDelegations);
-        }
-
-        public Task<HttpResponseMessage> RevokeReceivedMaskinportenScopeDelegation(string party, RevokeReceivedDelegation delegation)
-        {
-            ThrowExceptionIfTriggerParty(party);
-
-            IdValuePair resourceMatch = delegation.Rights.First().Resource.First();
-            IdValuePair fromMatch = delegation.From.First();
-
-            string path = Path.Combine(dataFolder, "MaskinportenSchema");
-            if (Directory.Exists(path))
-            {
-                string content = File.ReadAllText(Path.Combine(path, "backendReceived.json"));
-                List<MaskinportenSchemaDelegation> delegations = JsonSerializer.Deserialize<List<MaskinportenSchemaDelegation>>(content, options) ?? [];
-                foreach (MaskinportenSchemaDelegation d in delegations)
-                {
-                    if (d.CoveredByPartyId.ToString() == party &&
-                        d.OfferedByOrganizationNumber == fromMatch.Value &&
-                        d.ResourceId == resourceMatch.Value)
-                    {
-                        return Task.FromResult(new HttpResponseMessage(HttpStatusCode.NoContent));
-                    }
-                }
-            }
-
-            return Task.FromResult(new HttpResponseMessage(HttpStatusCode.InternalServerError));
-        }
-
-        public Task<HttpResponseMessage> RevokeOfferedMaskinportenScopeDelegation(string party, RevokeOfferedDelegation delegation)
-        {
-            ThrowExceptionIfTriggerParty(party);
-
-            IdValuePair resourceMatch = delegation.Rights.First().Resource.First();
-            IdValuePair toMatch = delegation.To.First();
-
-            string path = Path.Combine(dataFolder, "MaskinportenSchema");
-            if (Directory.Exists(path))
-            {
-                string content = File.ReadAllText(Path.Combine(path, "backendOffered.json"));
-
-                List<MaskinportenSchemaDelegation> delegations = JsonSerializer.Deserialize<List<MaskinportenSchemaDelegation>>(content, options) ?? [];
-                foreach (MaskinportenSchemaDelegation d in delegations)
-                {
-                    if (d.OfferedByPartyId.ToString() == party &&
-                        d.CoveredByOrganizationNumber == toMatch.Value &&
-                        d.ResourceId == resourceMatch.Value)
-                    {
-                        return Task.FromResult(new HttpResponseMessage(HttpStatusCode.NoContent));
-                    }
-                }
-            }
-
-            return Task.FromResult(new HttpResponseMessage(HttpStatusCode.InternalServerError));
-        }
-
-        public Task<HttpResponseMessage> CreateMaskinportenScopeDelegation(string party, DelegationInput delegation)
-        {
-            ThrowExceptionIfTriggerParty(party);
-
-            string resourceId = delegation.Rights.First().Resource.First().Value;
-            IdValuePair toMatch = delegation.To.First();
-            string path = Path.Combine(dataFolder, "MaskinportenSchema", "Delegation", $"{resourceId}.json");
-            if (File.Exists(path))
-            {
-                string content = File.ReadAllText(path);
-                return Task.FromResult(new HttpResponseMessage
-                { StatusCode = HttpStatusCode.Created, Content = new StringContent(content) });
-            }
-
-            return Task.FromResult(new HttpResponseMessage(HttpStatusCode.BadRequest));
-        }
-
-        /// <inheritdoc />
-        public Task<List<DelegationResponseData>> MaskinportenSchemaDelegationCheck(string partyId, Right request)
-        {
-            string resourceId = request.Resource[0].Value;
-            string filename;
-            switch (resourceId)
-            {
-                case "appid-400":
-                case "appid-136":
-                    filename = resourceId;
-                    break;
-                default:
-                    filename = "default";
-                    break;
-            }
-
-            string fullPath = Path.Combine(dataFolder, "MaskinportenSchema", "DelegationCheck", filename + ".json");
-
-            List<DelegationResponseData> mockedResponse = Util.GetMockData<List<DelegationResponseData>>(fullPath);
-
-            return Task.FromResult(mockedResponse);
-        }
-
-        //// SingleRight
-
-        public async Task<HttpResponseMessage> CheckSingleRightsDelegationAccess(string partyId, Right request)
-        {
-            ThrowExceptionIfTriggerParty(partyId);
-
-            string resourceFileName = GetMockDataFilenameFromUrn(request.Resource);
-            string path = Path.Combine(dataFolder, "SingleRight", "DelegationAccessCheckResponse");
-
-            return await Util.GetMockedHttpResponse(path, resourceFileName);
-        }
-
-        /// <inheritdoc />
-        public async Task<HttpResponseMessage> CreateSingleRightsDelegation(string party, DelegationInput delegation)
-        {
-            string resourceFileName = GetMockDataFilenameFromUrn(delegation.Rights.First().Resource);
-            string dataPath = Path.Combine(dataFolder, "SingleRight", "CreateDelegation");
-
-            return await Util.GetMockedHttpResponse(dataPath, resourceFileName);
-        }
-
-        /// <inheritdoc />
         public async Task<HttpResponseMessage> GetSingleRightsForRightholder(string party, string userId)
         {
 
@@ -365,24 +198,6 @@ namespace Altinn.AccessManagement.UI.Mocks.Mocks
             else
             {
                 return Task.FromResult(new HttpResponseMessage(HttpStatusCode.Created));
-            }
-        }
-
-        private static string GetMockDataFilenameFromUrn(List<IdValuePair> resourceReference)
-        {
-            IdValuePair referencePart = resourceReference.First();
-
-            switch (referencePart.Id)
-            {
-                case "urn:altinn:resource":
-                case "urn:altinn:servicecode":
-                case "urn:altinn:app":
-                    return referencePart.Value;
-                case "urn:altinn:org":
-                case "urn:altinn:serviceeditioncode":
-                    return resourceReference[1].Value;
-                default:
-                    return "Unknown";
             }
         }
