+++ backend/src/Altinn.AccessManagement.UI/Altinn.AccessManagement.UI/appsettings.Development.json
@@ -9,8 +9,16 @@
   },
   "GeneralSettings": {
     "FrontendBaseUrl": "https://am.ui.at22.altinn.cloud",
-    "Hostname": "at22.altinn.cloud",
-    "UseMockData": true
+    "Hostname": "at22.altinn.cloud"
+  },
+  "MockSettings": {
+    "AccessManagement_V0": false,
+    "AccessManagement": true,
+    "AccessPackage": true,
+    "ResourceRegistry": false,
+    "Profile": false,
+    "Register": false,
+    "KeyVault": false
   },
   "KeyVaultSettings": {
     "SecretUri": "https://altinn-dev-amui-kv.vault.azure.net/"
