+++ backend/src/Altinn.AccessManagement.UI/Altinn.AccessManagement.UI/appsettings.json
@@ -17,7 +17,15 @@
   "GeneralSettings": {
     "FrontendBaseUrl": "http://localhost:5101",
     "Hostname": "localhost",
-    "UseMockData": false
+  },
+  "MockSettings": {
+    "AccessManagement": false,
+    "AccessManagement_V0": false,
+    "AccessPackage": false,
+    "Profile": false,
+    "Register": false,
+    "ResourceRegistry": false,
+    "KeyVault": false
   },
   "KeyVaultSettings": {
     "SecretUri": ""
