+++ src/features/amUI/userRightsPage/DelegationModal/SingleRights/ResourceSearch.tsx
@@ -165,7 +165,7 @@ export const ResourceSearch = ({ onSelection, toParty }: ResourceSearchProps) =>
           avatar={{
             type: 'company',
             imageUrl: resource.resourceOwnerLogoUrl,
-            name: resource.resourceOwnerName,
+            name: resource.resourceOwnerName ?? '',
           }}
         />
         {!!delegatedResources &&
